퀴즈 타입 추가 및 퀴즈 컴포넌트 분리
