+++ src/components/quiz/QuizBox.jsx
@@ -6,6 +6,8 @@ import QuizFrame from "../common/QuizFrame";
 import Explanation from "./Explanation";
 import ConfirmChoice from "./ConfirmChoice";
 import { USERDATA } from "../common/key";
+import MultipleQuiz from "./MultipleQuiz";
+import OxQuiz from "./OxQuiz";
 
 export default function QuizBox() {
   const {
@@ -18,70 +20,27 @@ export default function QuizBox() {
     showModal,
     setShowModal,
     showExplanation,
-    userCount,
-    explanation,
-    userAnswer,
-    answer,
+    updateAnswerArr,
+    quizType,
+    setQuizType,
   } = useContext(QuizContext);
-  const [answerArr, updateAnswerArr] = useImmer([]);
-  const [selectedAns, setSlectedAns] = useState("false");
 
-  const answers = useRef([]);
   const userData = JSON.parse(localStorage.getItem(USERDATA));
 
   useEffect(() => {
-    GetQuizQuestion(setUserCount, setQuestion, updateAnswerArr, setExplanation, setAnswer, userData);
+    GetQuizQuestion(setUserCount, setQuestion, updateAnswerArr, setExplanation, setAnswer, userData, setQuizType);
   }, []);
 
-  function select(n) {
-    for (let i = 0; i < 4; i++) {
-      if (i == n) {
-        answers.current[i].style.setProperty("border", "3px solid #FF9748");
-      } else {
-        answers.current[i].style.setProperty("border", "3px solid #ffd1af");
-      }
-    }
-  }
   return (
     <>
       {showModal && <ConfirmChoice />}
-      {!showExplanation && (
-        <QuizFrame>
-          <Question>{`Q.  ${question} 이것에 대한 정답은 무엇일까요?`}</Question>
-          <AnswerBox>
-            {answerArr.map((a, n) => (
-              <Answer
-                ref={(el) => (answers.current[n] = el)}
-                onClick={() => {
-                  select(n);
-                  updateUserAnswer((obj) => {
-                    obj.answer = a.answer;
-                    obj.result = a.result;
-                  });
-                  setSlectedAns("true");
-                }}
-                key={n}
-              >
-                {a.answer}
-              </Answer>
-            ))}
-          </AnswerBox>
-          <SelectButton
-            $slectedState={selectedAns}
-            onClick={() => {
-              setShowModal(true);
-            }}
-          >
-            {selectedAns == "false" ? "답을 선택해주세요" : "선택했어요"}
-          </SelectButton>
-        </QuizFrame>
-      )}
-      {showExplanation && <Explanation />}
+      {!showExplanation ? quizType == "m" ? <MultipleQuiz /> : <OxQuiz /> : <Explanation />}
     </>
   );
 }
-async function GetQuizQuestion(userCount, question, answerArr, explanation, answer, userData) {
+async function GetQuizQuestion(userCount, question, answerArr, explanation, answer, userData, setQuizType) {
   const response = await fetch(`http://localhost:8080/main/quiz/${userData.id}`);
+  // const response = await fetch("json/quiz.json");
   const data = await response.json();
 
   userCount(data.user.count);
@@ -98,6 +57,9 @@ async function GetQuizQuestion(userCount, question, answerArr, explanation, answ
       answer(data.quiz.example[e]);
       obj.result = "true";
     } else {
+      if (data.quiz.example[e] == null) {
+        setQuizType("o");
+      }
       obj.result = "false";
     }
     answerArr((arr) => {
@@ -106,45 +68,5 @@ async function GetQuizQuestion(userCount, question, answerArr, explanation, answ
       }
     });
   }
-
-  const Question = styled.p`
-    width: 80%;
-    font-size: 40px;
-  `;
-
-  const AnswerBox = styled.section`
-    width: 80%;
-    height: 40%;
-    display: flex;
-    flex-wrap: wrap;
-    align-items: center;
-    justify-content: space-evenly;
-  `;
-
-  const Answer = styled.p`
-    width: 45%;
-    height: 45%;
-    border-radius: 20px;
-    border: 3px solid #ffd1af;
-    font-size: 23px;
-    font-weight: 900;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    background: #ffeada;
-  `;
-
-  const SelectButton = styled.button`
-    width: 30%;
-    height: 12%;
-    border-radius: 20px;
-    border: 3px solid ${(props) => (props.$slectedState == "false" ? "#DBD5D0" : "#ffbc89")};
-    color: ${(props) => (props.$slectedState == "false" ? "#6D6D6D" : "#f9f9f9")};
-    font-size: 20px;
-    font-weight: 900;
-    background-color: ${(props) => (props.$slectedState == "false" ? "#DBD5D0" : "#ff9748")};
-    outline: none;
-  `;
-
   explanation(data.quiz.explanation);
 }
